재고 업데이트 및 헤더 수정, 결제 화면 일부 구현

+++ src/main/java/com/spring/myHouse/inventory/repository/InventoryRepository.java
@@ -19,5 +19,5 @@ public interface InventoryRepository extends JpaRepository<Inventory, Long> {
     @Modifying
     @Transactional
     @Query("UPDATE Inventory i SET i.inventorycount = :inventorycount  WHERE i.productnum = :productnum")
-    Long updateInventoryCount(Long productnum, Long inventorycount);
+    int updateInventoryCount(Long productnum, Long inventorycount);
 }
